Move RoadPlacementArgs to CreateRoad.h
